feat: add auth for login and signup
