+++ src/AuthorizationPage/SignUp/Sign-up.css
