+++ src/AuthorizationPage/SignUp/Sign-up.scss
