+++ src/AuthorizationPage/SignUp/Sign-up.scss.d.ts
