+++ src/ErrorPage/Error.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import {useNavigate} from "react-router-dom";
-import Button from '../UI/Button/Button';
+import Button from '../components/UI/Button/Button';
 
 const Error = () => {
   const navigate = useNavigate();
