+++ src/UI/Footer/Footer.scss.d.ts
