+++ src/components/hok/RequireAuth.tsx
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react';
 import { Navigate, useLocation, useNavigate} from "react-router-dom";
 import { useSelector } from 'react-redux';
 import { RootState } from '../../store';
-import MainPage from '../MainPage/MainPage';
+import MainPage from '../../MainPage/MainPage';
 
  const RequireAuth = () => {
   const { isAuth } = useSelector((state: RootState) => state.signIn);
